Update music_generator.py
Added a second option for the mutation function, as well as I contained the MIDI file generation inside of a function so it is a bit less just floating in main. Lastly, you will most likely need to remove "Music_Generator_GA" in front of the melody in the imports section. I have to have this in my code due to my project file structure. Sorry for the inconvenience on that guys